Creating class ROILabel Change class screen
